i split main from criss cross
